pyproject.toml
[project]
name = "create-drawio"
version = "0.1.0"
description = "Python File for Creating ERD drawio file"
authors = [
    {name = "Mauro Cerzosimo",email = "cerzosimo@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "watchdog (>=6.0.0,<7.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
]

[tool.poetry]
packages = [{include = "drawio_tools", from = "src"}]


[tool.poetry.group.dev.dependencies]
ruff = "^0.11.8"
black = "^25.1.0"
pre-commit = "^4.2.0"
jupyterlab = "^4.4.1"
mypy = "^1.15.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
select = ["E", "F", "B"]  # E = pycodestyle, F = pyflakes, B = bugbear

[tool.black]
line-length = 88

[tool.poetry-plugin-postinstall]
hook = "pre-commit install"
